feat: Add Medusa Admin API resources with enum-based type safety
✨ Features Added:
• Complete Sellers API resource with CRUD operations
• Requests API resource for managing seller/product requests
• Commission/Commission-lines API resources for financial tracking

📦 New Models:
• Seller, SellerInvitation, SellerMember models
• Request model for handling approval workflows
• Commission, CommissionLine models for financial tracking

🔧 API Clients (Retrofit):
• AdminSellersResource - Full CRUD + specialized endpoints
• AdminRequestsResource - List, retrieve, review requests
• AdminCommissionResource - Commission lines management

📋 Params/Responses/Requests:
• AdminGetSellersParams, AdminGetRequestsParams, AdminGetCommissionLinesParams
• AdminSellersRes, AdminRequestsRes, AdminCommissionLinesRes, etc.
• AdminUpdateSellerReq, AdminInviteSellerReq, AdminReviewRequestReq

🎯 Enum-based Type Safety:
• StoreStatus (active, inactive, suspended)
• RequestType (product, seller, product_collection, etc.)
• RequestStatus (pending, accepted, rejected)
• CommissionStatus (pending, paid, cancelled)
• CommissionType (commission, fee, bonus, penalty)

🔄 Refactored Models:
• All string-based status fields converted to enums
• Added @JsonEnum(fieldRename: FieldRename.snake) annotations
• Updated all params/requests to use enums where appropriate

📄 Documentation:
• Added WARP.md with comprehensive implementation guide
• All new files properly exported in index.dart files
• Generated code with build_runner for .g.dart files

Following WARP.md guidelines for structured API client generation.

+++ lib/src/models/inventory_item.dart
@@ -49,13 +49,13 @@ class InventoryItem {
 
   final String? material;
 
-  final int? weight;
+  final double? weight;
 
-  final int? length;
+  final double? length;
 
-  final int? height;
+  final double? height;
 
-  final int? width;
+  final double? width;
 
   final bool requiresShipping;
